pyproject.toml
[tool.poetry]
name = "watermarkbot"
version = "1.5.3"
description = ""
authors = ["Divanshu Chauhan <23524935+Divkix@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Pyrogram = "^2.0.26"
hachoir = "^3.1.3"
aiofiles = "^0.8.0"
aiohttp = "^3.8.1"
dnspython = "^2.2.1"
humanfriendly = "^10.0"
psutil = "^5.9.1"
TgCrypto = "^1.2.3"
cachetools = "^5.2.0"
pymongo = "^4.1.1"
Pillow = "^9.1.1"
prettyconf = "^2.2.1"
pickledb-ujson = "^1.0.6"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"
pre-commit = "^2.19.0"
pyclean = "^2.2.0"
pylint = "^2.13.9"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
